lib: remove `impl Trait` return and inline for concrete types
Apparently Rust 2024 has made a change to lifetime inference of `impl
Trait`s, such that this code for property testing is now invalid --
the lifetime of the input reference 'state' in testutils.rs (which
calls arb_* functions) is inferred differently, such that there is no
guarantee the returned value has a lifetime exceeding that of the input
reference.

However, this can simply be side-stepped by just inlining the two uses
of the generalized `impl Trait` return types into their callers, and
giving them concrete types.

Signed-off-by: Austin Seipp <aseipp@pobox.com>